speaking_service/requirements.txt
openai>=1.0
python-dotenv
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6